Update patch set 3
Patch Set 3:

(2 comments)

Hi
This patch will be reduced to the introduction section only. 
I think it will be good to build a basis from which we can start. I think some of the other sections will require a bit of rework which is why I've broken it down into sections.

Patch-set: 3
Label: Verified=0