trafficgens: Add 'trafficgen' implementations
There are three trafficgen implementations supported:

* Ixia (IXIA IxExplorer)
* IxNet (IXIA IxNetwork)
* Dummy

Not all of these trafficgen implementations provide the entire list of
features defined by the 'trafficgen' interface. This is because either
the features are not available for the produce (latency measurements
for the 'Ixia' trafficgen) or are not yet required by tests (burst
traffic for the 'IxNet' trafficgen). However, these features (plus
additional trafficgen implementations) can be added in time.

Change-Id: Icfd46255d117c407d5b99c67156def6428f3c115
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
JIRA: VSPERF-27