Update patch set 4
Patch Set 4:

> The change looks good to me but I don't understand why the build is
 > failing with this error:
 > 
 > checking whether make has GNU make $(if) extension... yes
 > configure: error: cannot link with dpdk

Not sure TBH, it's a little odd because we do modify dpdk to build the combined library and you should only see that if the combined library isn't building.
I tried the latest patch locally and it seems to be working. Please let me know if you can test it and if you are having any issues.

thanks
Maryam

Patch-set: 4
Label: Verified=0