Update patch set 2
Patch Set 2:

Maryam, the original dpdk.conf inside jobs was meant to give an example that different provider may put their own list of configuration choices, for example to use ovs-system or ovs-dpdk etc. The example file probably should be better named intel.conf, which will pick ovs-dpdk as the switch choice. It was not meant to configure the dpdk software. Since I see you are moving a python file here, I would assume your python program was to configure the dpdk program. If that is the case, it should belong to dpdk specific directory.

Patch-set: 2