Update patch set 3
Patch Set 3:

The PDV uses the minimum delay as the reference, but the minimum delay itself could be a outlier. This makes PDV result unreliable. 

Why not just take a mean and standard variations from all the delay measurement? Statistics has already provided the strict model for this kind of problems.

Patch-set: 3