Update patch set 5
Patch Set 5:

I put a comment inside body, but the comment did not seem to get published - it is still shown as draft. I would like to add "or equal to" in the following sentence.

The average latency for traffic traversing the slow-path should be greater than or equal to the latency of traffic traversing the fastpath.

Patch-set: 5