Add a simple performance test that sends a continuous stream
Implement a simple performance test that sends a continuous
stream of traffic to the vSwitch at line rate for 30 seconds
and measures the throughput
Note: there is issue with calculating of the Tx Fps and Mbps
based on the results from the IxNetwork traffic generator.
This issue will be fixed in a later patch.

JIRA: VSPERF-27

Change-Id: Ic72225888735d33398a602462af159e901fb5957
Signed-off-by: Radek Zetik <radekx.zetik@intel.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>