bug_fix: process pkt_size/duration from cli
Refactor rfc25444_duration to duration and update the --test-params help
to show the type of parameters that can be set from the cli. Also
configure pkt_sizes so that it can be modified from the CLI. Also fixes
the tcl script to configure framesize properly (to the value selected
and not the default 64b).

JIRA: VSPERF-27

Change-Id: I76dcd824977a5f940cf7476090d2d2a3acca8160
Signed-off-by: Billy O'Mahony<billy.o.mahony@intel.com>
Signed-off-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Al Morton <acmorton@att.com>
Reviewed-by: Dino Madarang <dino.simeonx.madarang@intel.com>