Update patch set 3
Patch Set 3:

This patch has a dependency on the other patch to clone and build qemu, is this correct?
I've tried to verify it but I don't think it's possible without that?
if yes we should build the correct dependency chain here or call out what we depend on

Patch-set: 3