Update patch set 1
Patch Set 1:

Hi Martin,

the positional args functionality was lost moving from the old toit code to vsperf. ie the -h output was out of date for the and there was no code there to implement it. 

So when I came to adding this of functionality I started from scratch.

Patch-set: 1
Label: Verified=0