test_spec: LTD: Some "Throughput" Tests belong in another category
In the LTD, some of the tests with "Throughput" in the title are not actually
testing throughput in the data path.
Since the test names are used by TOIT/VSPERF to group similar tests for batch
execution, it's important that the names accurately represent the test methods
and metrics measured.

JIRA: VSPERF-98

Change-Id: I620ddc0779232f2dd325379005faba8692fea18c
Signed-off-by: Al Morton <acmorton@att.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Billy O'Mahony<billy.o.mahony@intel.com>
Reviewed-by: Gene Snider <eugene.snider@huawei.com>