test_spec: LTD:Throughput.RFC2544.Profile Metric Clarifications
LTD.Throughput.RFC2544.Profile refers to the metric "maximum forwarding rate"
but implies that this is the zero% loss rate, which would be the Throughput
level (MFR is usually higher than Throughput, because MFR allows losses).
It may be that MFR and Throughput are just interchanged in the last two
paragraphs of the description.

JIRA: VSPERF-110

Change-Id: I4f36fd4ce66f2f4ebee4f60a11ad03adfb2b3b7c
Signed-off-by: Al Morton <acmorton@att.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Billy O'Mahony<billy.o.mahony@intel.com>
Reviewed-by: Gene Snider <eugene.snider@huawei.com>