Update patch set 2
Patch Set 2:

Hi Billy,  When we ask the question, "When we make the traffic exercise much more of the fastpath cache (e.g. by all available udp port numbers in the packets) THEN what is the max throughput?" part of the answer (I believe) is knowing the delta from the "normal" 2544 Throughput test. So, I'm suggesting that 2544 Throughput needs to be run before this one, even if there isn't a procedural dependency.  OK?

Patch-set: 2
Label: Verified=0