Update patch set 4
Patch Set 4:

Hi Al,

sorry I didn't get a chance to look at gerrit for a while after my initial comment.

I agree with you if pre-req is in the sytle of referring back to a delta to the other test then that is fine. I just didn't want to go back to the previous statement of the test which was equivalent but quite a bit more complicated to implement!

Patch-set: 4
Reviewer: Gerrit User 679 <679@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0