Update patch set 2
Patch Set 2:

Billy, I agree with the specification transition to Genome-specs over at Functest, They have some variable ranges in their percentage traffic specs, and that would cause problems with comparisons.

Also: ":) +1 on that Brian. We also have no interest in following that path so thanks for pointing out (when all you have is a hammer everything look looks like a nail as they say so when all you have is an Ixia...)"
>>> you pound nails with it? No, but this explains what happened very well...

Patch-set: 2