Update patch set 2
Patch Set 2:

Hi Gene, I'm thinking that another solution to produce identical IMIXes in different packet generators would be to have a common API to the custom IMIX configuration capability of each system. We need this in the future anyway, to support the custom IMIXes that will come from Functest Traffic Profiles. My guess is that none of the "named" distributions will line-up with these Profiles.  /Al

Patch-set: 2