Update patch set 2
Patch Set 2:

> @Maryam. Not sure what looks good to you: The code or my proposed
 > changes. 

What was the proposed change? I think I missed that comment/can't see it.

 > I hope it's the latter. :-) There is some overlap between
 > the canned iMix settings for Ixia and Spirent. Both tools have the
 > capability to support a custom mix, so translation between the
 > tools can be done. I must object on principle to a vendor-neutral
 > framework coded to support one vendor's way of doing things. My
 > proposed fix allows a consumer using Ixia to use Ixia iMix names,
 > and a consumer using Spirent to use Spirent iMix names. It also
 > allows the consumer to specify a custom mix in the spirit of the
 > iMix Genome: https://tools.ietf.org/html/draft-morton-bmwg-imix-genome-02.

I'm for vendor neutrality, which is why I'd like to understand as a starting point what overlap in canned IMIXes there are between the various traffic gens. We can look at specifying and implementing custom IMIXes in the long term but it's not realistic for release B.
Even if it was to select 1 standard IMIX today, it would be good to add this basic functionality as a start.
Other option is we leave it out for release B until we define a custom IMIX using the IMIX genome methodology.

Patch-set: 2