Update patch set 2
Patch Set 2:

Maryam wrote: "I'm for vendor neutrality, which is why I'd like to understand as a starting point what overlap in canned IMIXes there are between the various traffic gens. We can look at specifying and implementing custom IMIXes in the long term but it's not realistic for release B. Even if it was to select 1 standard IMIX today, it would be good to add this basic functionality as a start. Other option is we leave it out for release B until we define a custom IMIX using the IMIX genome methodology."

The simplest form of my proposed change is that the tool accept an arbitrary iMix string that is to be interpreted by the traffic gen. If that string is "default", the TG uses it's default iMix. If that string is "cisco" and the TG is Ixia, the Ixia TG knows what to do. If that string is "JMIX Downstream" and the TG is Spirent, the Spirent TG knows what to do. Such a proposal would be rapid to implement for vsperf using only "default". It also puts the responisbility on the TGs to support their individual mixes. It would give the tester the flexibility to run whatever iMix flavor the TG of choice has to offer. We can postpone custom mixes/genome support until later.

Patch-set: 2