Update patch set 4
Patch Set 4:

Build failures are seen during build of 2nd and 3rd copy of Qemu. This functionality has been implemented by VSPERF-85 https://gerrit.opnfv.org/gerrit/#/c/2207/. Since this feature has been merged, build failures are observed at 2nd (src_cuse) or 3rd (src_vanilla) Qemu build. Occurrence is random and it has started before VSPERF-95 was event sent for the review. It means, that VSPERF-95 is not the cause, but one of the victims of build failures.

It is to note, that there is a call to configure in current makefile, but despite that build is sometimes failing. Radek (author of original patch) and Fatih were investigating this issue, but the root cause has not been found yet. It has been agreed, that it is not needed to build Qemu three times, because there aren

Patch-set: 4
Reviewer: Gerrit User 1306 <1306@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0