Update patch set 3
Patch Set 3:

I think this only passed by accident. See my comments on 2735. I will check with RelEng and make sure they delete our workspace for each build so that we get a clean environment each time. That's the only way we can be sure it works "out of the box". It will also make sure we are testing with the current version and not whatever happens to be present in the tree when the job runs.

Patch-set: 3