Results: Integrate with opnfv_test_dashboard
All the test projects generate results in different format.
The goal of a testing dashboard is to provide a consistent
view of the different tests from the different projects.
We need to enable the data collection for VSPERF, in other words
VSPERF pushes data using a REST API.
The feature is enabled by --opnfvpod parameter on vsperf command line.
The value of the parameter sets POD name.
Example:
--opnfvpod <pod_name>   or   --opnfvpod=<pod_name>
You need to specify installer name. It can be set in conf-file
default value:
OPNFV_INSTALLER = "Fuel"

JIRA: VSPERF-112

Change-Id: I37255414eebcc2ff0c12109c508dcd984663a830
Signed-off-by: Radek Zetik <radekx.zetik@intel.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Brian Castelli <brian.castelli@spirent.com>