Update patch set 1
Patch Set 1:

(1 comment)

I'm not going to block it but you really, really, really,really want to specify the exact versions to install but adding it to the package name. At Huawei we have had several issues when packages, particularly python setuptools, released a new, incompatible version. You can probably do this simply by adding '=1.2.3.4' to the package name ( with the correct version of course). There are also issues when one package has been updated and has new dependencies but the required packages have not released the new version yet. And so on, and so on and so on.

Patch-set: 1