Update patch set 1
Patch Set 1:

I'm a bit confused here. pip is installing all python packages from requirements.txt and the package names are there set with specific versions('==1.2.3.4') like Gene mentioned. The are same for all system build scripts anyway and aren't changed based on distros. 
We are facing issues with different versions of python being picked if we rely on default python/pip depending on the Linux distro, as a results some packages are not being installed/missing. That was the reason for using version specific virtualenv-3.x to ensure that we have a python3 environment and required packages are installed in a virtual environment. Maybe I'm still missing something?

Thanks,
Abdul

Patch-set: 1