docs: updates and move traffic gens to separate doc
Move the traffic gen instructions to a separate user guide and add
information on usage of the Dummy traffic generator. Update docs
to fix PDF build failure and do general clean-up. Removed the numbering
from the LTD and added the numbered directive to automate numbering for
sections and headers. Add comment anchors that reflect the section
numbers.

Change-Id: I984ca38456a891c439697ebc1da041bc1d828a15
Signed-off-by: Maryam Tahhan <maryam.tahhan@intel.com>