Update patch set 31
Patch Set 31:

> It seems that when you have a title, you must also have a
 > sub-title:
 > 
 > ===============
 > Title
 > ===============
 > ---------------
 > Sub-title
 > ---------------
 > 
 > 1.0 Introduction
 > ------------------
 > this is the intro
 > 
 > 1.1 Sub-section
 > :::::::::::::::
 > more text
 > 
 > 1.1.1 sub sub-section
 > ~~~~~~~~~~~~~~~~~~~~~
 > 
 > 1.1.1.1 sub sub sub-section
 > ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 > 
 > All of this converts fine at http://rst.ninjs.org/

Hi Al
I was trying to follow this heading structure as well... the issue is that the TOC was set too deep, so the PDF generation was failing - I've fixed this up now so the titles of the DOCs are moved out to index.rst, so we never get too deep for the PDF...

One last problem - we're going to have to get rid of the numbering in the section headers I think. so just using the :numbered: directive to look after the numbering for us. the issue is in the generated PDF - we inherit the :numbered: from the top level index.rst which causes some funky numbering to happen...

Patch-set: 31