Update patch set 3
Patch Set 3:

> Hi Maryam,  The frame sizes we specify in the LTD are the standard
 > Ethernet set:
 > Frame size (bytes): 64, 128, 256, 512, 1024, 1280, 1518, 2K, 4k
 > 
 > So, as long as results[0] is also reported at Ethernet layer (and
 > not at IP), the units and layers match, and we're good to go.
 > 
 > This may be a point that's more relevant for the new traffic
 > generators, but even 2544 is a bit ambiguous about frames or
 > packets sometimes...

At the moment they are at the Ethernet layer yes, I think it's worth adding a check though to read the framesize used from the test and check that what was used is indeed what the vsperf user selected.
for example: in some cases a crc is added to the end of a packet, resulting in 68bytes being sent over the wire instead of 64. --> Sounds like a JIRA ticket :)

Patch-set: 3