Update patch set 2
Patch Set 2:

> (15 comments)
 > 
 > Trying to do too much in the patch...
 > back2back should be a separate patch - 1 logical change.
 > I don't understand/see where the rest API gets invoked? Also if
 > there is configuration that is shared in back to back and
 > throughput test can you retrieve the settings in a separate
 > function rather than copy paste

Maryam,
The file testcenter-rfc2544-rest.py is independent of vsperf. The only file that is integrated with vsperf if testcenter.py. That is the testcenter.py is reading configuration file and invoking appropriate script. Hence, the testcenter-rfc2544-rest.py has main and not wrapped in the class. So, most of your comments on this file is related to this.
IN testcenter.py you'll find subprocess command where the appropriate script is called.

Christian and Maryam, 

Thanks a lot. I'll make other changes regarding uncommented code, appropriate logging-level,  etc.

Patch-set: 2