Update patch set 4
Patch Set 4:

Hi Sridhar
Comments inline
 > > (15 comments)
 > >
 > > Trying to do too much in the patch...
 > > back2back should be a separate patch - 1 logical change.
 > > I don't understand/see where the rest API gets invoked? Also if
 > > there is configuration that is shared in back to back and
 > > throughput test can you retrieve the settings in a separate
 > > function rather than copy paste
 > 

Still applies, please make sure to fix this.

 > Maryam,
 > The file testcenter-rfc2544-rest.py is independent of vsperf. The
 > only file that is integrated with vsperf if testcenter.py. That is
 > the testcenter.py is reading configuration file and invoking
 > appropriate script. Hence, the testcenter-rfc2544-rest.py has main
 > and not wrapped in the class. So, most of your comments on this
 > file is related to this.
 > IN testcenter.py you'll find subprocess command where the
 > appropriate script is called.
 > 

I understand this is the way it was done in the past. so I will merge it in (when the changes I mentioned above are met). However, this is far from the way it should be used with python. I would like to see some serious cleanup here in the future. this is not the way the api is intended to work. Please see the ixia directory for more information. We have should have 1 testcenter class and the other python files should be wrapped up in classes as well. the overlap in setting and arg parsing is too heavy. we should use the power of python to our advantage and not code in a c fashion.

Thanks for your patience.

Patch-set: 4