Update patch set 3
Patch Set 3:

To be honest I was thinking about the same, i.e. to add an optional param "VSWITCH" and "VNF" to the testcase definition, which will override default VSWITCH and VNF settings. I was not sure about TRAFFICGEN specific TCs, but it might also make sense if some TRAFFICGEN specific features will be used (or some trafficgen implementations will lack some generic features).
Finally I decided to postpone it and think about it first.

If you will create a JIRA ticket and/or start a discussion, it will be great.

Patch-set: 3