Update patch set 3
Patch Set 3:

Thanks Maryam,

Will address the copyright issues.

On the IP address in the baseconfig.x2544 file I will just remove the IP address as its not needed. This file is only used as the base config. The xena_json.py/xena.py->_setup_json_config is designed to take this config and modify based on the configuration in the 10_custom.conf/traffic profile and write out a new file called 2bUsed.x2544. This new 2bUsed.x2544 is the configuration that ends up being used when running the Xena2544.exe executable.

Patch-set: 3
Label: Verified=0