systems: Fix build scripts from arch specific packages
Fixes the build scripts to stop using arch specific packages.
The issue was found on an x86_64 arch when performing the
installation script. The glibc.i686 would fail to install
because a dependency was installed with a x86_64 arch.

While testing I found other versions of Fedora should also
include the glib2-devel package which was added to Fedora 22
build_base_machine script in another patch. This does not
seem to be included in the base installation of the OS.

Tested on Centos7, Fedora22, 21, 20, and Rhel7.2.

JIRA: VSPERF-281

Change-Id: I884d14fdacf0ceb96c9dc341e82c96d864653766
Signed-off-by: Christian Trautman <ctrautma@redhat.com>