Update patch set 1
Patch Set 1:

(2 comments)

The reason I suggested this change was I saw two issues on the code. One was in the Xena code where it was assuming a type of Boolean would be passed in the traffic profile for this parameter to setup either mesh or west-east config (half/full) duplex traffic. I assumed this based on the traffic defaults structure. I fixed this in b2b patch for Xena to typecast the value to Boolean in my if statement. I will change it to do a string compare based on my tests for my below statements to Martin.

Then as I looked into the code I found the issues in the vswitch_controller_pvp/pvvp switch setup modules. I decided it would be better to make things consistent so 

1. We try to create enforcement and a constant standard for the value
2. Prevent future bugs such as mine when creating traffic generator modules that are using the traffic defaults dictionary to see what type of values are going to be passed and how those values are passed.

If Ixia requires it to be True/False instead of true/false I will update this patch to reflect that. I can use the title method on the string instead of lowercase to keep it consistent from the testcases conf files.

Patch-set: 1
Reviewer: Gerrit User 2851 <2851@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0