linux_bridge_guest: Modify packet structure for linux bridge tests
Change packet structure for pvp/pvvp tests that use the linux
bridge as the loopback to include the tgen mac address instead
of the guest nic mac addresses in layer 2.

This change is due to the old way not working with Xena traffic
generators. Ixia works with both the old and new way.

 * Verified to work with Xena
 * Verified to work with Ixia

JIRA: VSPERF-307

Change-Id: Ieca0c58610797d9d5c726ba9f21227e255958092
Signed-off-by: Christian Trautman <ctrautma@redhat.com>