Namespace_veth: Add funtionality for network namespace, veth ports
Adds functionality for network namespaces and veth/peer ports
to connect the namespaces. The namespaces are tracked outside
of default Linux namespaces to prevent possible disruption of
work outside of VSPerf where a user may create a namespace and
the cleanup code deletes it. The cleanup code only removes
namespaces and veth ports created inside the testcase.

JIRA: VSPERF-310

Change-Id: If8881cafb119f38f052403a1de497e9660187d2e
Signed-off-by: Christian Trautman <ctrautma@redhat.com>