Update patch set 2
Patch Set 2:

Hi Bill,

I could write something to monitor the output of a command to see when rstp/stp is enabled/disabled, but there's no easy way to not clutter up the log with a lot of output checking the bridge info command. Since this is going to be used for integration tests, the validation step would fail if there was an issue, so I'm going to leave the sleeps in place.

Sridhar I fixed the imports, because there were methods in this module that had sleep as a param name I decided to use the time namespace call on my sleep calls and removed the import. I did fix the import ordering.

Patch-set: 2