Update patch set 3
Patch Set 3:

Martin,
in yardstick.rst, 
Apart from minor typos such as comma before that, missing fullstop,  etc, I was able to understand the document easily.
I have just comment: Can we name the 2nd option as traffic generator mode? as that is how it is referred in other part of the document.

Patch-set: 3