xena_deadlock: Fix xena2544.exe deadlock
Fix Xena2544.exe execution with mono where sometimes the
Xena process says complete, but the mono process doesn't exit.

Changes call of Xena2544.exe from communicate thread lock to a
wait try/catch block and checks the log periodically to see
if a deadlock has occured and if so, the mono pid is terminated
to prevent deadlock.

* Fix one small issue with typecasting to int, changed to float

JIRA: VSPERF-319

Change-Id: Ie7a86a2b9ff9cdfacba5a912ad17df6cc0581702
Signed-off-by: Christian Trautman <ctrautma@redhat.com>