Update patch set 2
Patch Set 2:

(2 comments)

Hi Mars,

Just one thought and one thing we should probably update in the docs for this change.

With this implementation we should probably update the vswitchperf/docs/configguide/installation.rst hugepage section to state that we will allocate hugepages if none are configured based on this new setting.

Patch-set: 2