pkt_gen: MoonGen start rates not correct
The VSPERF traffic profile for starting rate is in a
percentage of line rate.  MoonGen programs its
starting rate as Mpps.  Therefore the traffic profile
for VSPERF needs to be converted to a MoonGen format.
There was a hardcoding of 14.88 that was used for the
starting frame rate for all frame sizes.  This needs
to be corrected to take into account all frame sizes.

JIRA: VSPERF-350

Change-Id: Ied71668a560d7d2b03483f9a1192dd2fb4965996
Signed-off-by: bmichalo <bmichalo@redhat.com>