integration_cleanup: Force testcase cleanup to always run
During testcase development on the integration side, if
namespaces are used or veth ports and an exception is raised.
The namespaces and veth ports are not properly cleanup up
because the run_finalize does not get executed.

This patch wraps the testing block in a try block where
a finally will ALWAYS execute the run_finalize thus
preventing the possiblity of leftover testing artifacts.

JIRA: VSPERF-328

Change-Id: I4ea5f29a48b9171f330847f34b5b30ef3ab22f27
Signed-off-by: Christian Trautman <ctrautma@redhat.com>