CLI: add Multiqueue params.
This patch implements the following 4 new CLI parameters that
can be used in the --test-params list.

 1. guest_nic_queues, which overrides GUEST_NIC_QUEUES
 2. guest_testpmd_txq, which overrides GUEST_TESTPMD_TXQ
 3. guest_testpmd_rxq, which overrides GUEST_TESTPMD_RXQ
 4. guest_testpmd_nb_cores, which overrides GUEST_TESTPMD_NB_CORES

When they are present in the CLI they will override the
corresponding existing settings in the configuration files.

As an example a usage can be:
./vsperf --test-params "guest_nic_queues=2;guest_testpmd_txq=1;\
guest_testpmd_rxq=1;guest_testpmd_nb_cores=1"...

JIRA: VSPERF-371

Change-Id: I140d6e068c1653d21aaa0df27532a48c4a12c151
Signed-off-by: Antonio Fischetti <antonio.fischetti@intel.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Martin Klozik <martinx.klozik@intel.com>