CLI: add Multiqueue params.
This patch implements the following new CLI parameters that
can be used in the --test-params list.

 1. guest_nic_queues, which overrides all GUEST_NIC_QUEUES values
 2. guest_testpmd_txq, which overrides all GUEST_TESTPMD_TXQ
 3. guest_testpmd_rxq, which overrides all GUEST_TESTPMD_RXQ
 4. guest_testpmd_nb_cores, which overrides all GUEST_TESTPMD_NB_CORES
    values
 5. guest_testpmd_cpu_mask, which overrides all GUEST_TESTPMD_CPU_MASK
    values
 6. vswitch_dpdk_multi_queues, which overrides VSWITCH_DPDK_MULTI_QUEUES
 7. guest_smp, which overrides all GUEST_SMP values
 8. guest_core_binding, which overrides all GUEST_CORE_BINDING values

When they are present in the CLI the corresponding settings from the
configuration files will be overridden.

As an example a usage can be:
./vsperf --test-params "guest_nic_queues=2;guest_testpmd_txq=1;\
guest_testpmd_rxq=1;guest_testpmd_nb_cores=1"...

v2: rebase + rework to add other CLI params as suggested and update doc.
v3: changes in 'testusage.rst' doc.

JIRA: VSPERF-371

Change-Id: I140d6e068c1653d21aaa0df27532a48c4a12c151
Signed-off-by: Antonio Fischetti <antonio.fischetti@intel.com>
Reviewed-by: Maryam Tahhan <maryam.tahhan@intel.com>
Reviewed-by: Martin Klozik <martinx.klozik@intel.com>