Update patch set 1
Patch Set 1:

(4 comments)

This is a lot of code so I'm try to run it through some tests this weekend. Its a lot of things to test so I'll get as much done as possible. I don't see anything glaring that jumps out for issues in the coding as it looks very well thought out. I just have a few grammar comments for the documentation and one thing I noticed in the 04_vnf.conf file.

This will be a fun rebase for whoever gets their patch merged first :).

Patch-set: 1
Label: Code-Review=0