Update patch set 1
Patch Set 1:

Hi Alex,

welcome aboard! I was bordered for some time, that we are so closely tight to the x86_64 platform. So again, it's great, that you're doing some investigation in this direction. I'm not sure how far you're with running VSPERF at different platform. In case it is a work in progress, then I would suggest to get it up and running first. It is possible (and very likely), that more complex changes will be needed. Once you'll know what's needed, it would be easier to discuss it and find an efficient solution.

We should also create a jira ticket to track multiarch support and link it from the patch(es). As discussed before, it would be good to merge bits and pieces of multiarch support into a bigger patch, which will bring a real functionality.

You could always reach me via mail, if you will have any question.

Best Regards,
Martin

Patch-set: 1