Update patch set 4
Patch Set 4:

(6 comments)

Looks great overall.
One general comment, we should report out some of the test configuration as part of the results also. I don't mind if this is in a separate patch.
But part of the test report should now reflect the # of VMs the # of Guest NICs, the type of Guest NICs (vhost/other), the configuration (parallel or serial), the affinitization...  the # of VMs + configuration could probably be inferred from the test name (in fairness), but I think the NICs comment and the affinitization still apply

Patch-set: 4
Reviewer: Gerrit User 304 <304@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0