Update patch set 4
Patch Set 4:

I plan to retest my changes with your patch
(https://gerrit.opnfv.org/gerrit/#/c/20015/ which is already merged) and if I
encounter any error, I will fix it.

If there is no problem with usage of paths in vsperf when running it outside
its own directory, I will keep this patch as it is.

Patch-set: 4