Update patch set 3
Patch Set 3:

Hi Sridhar,

do you plan to do address also my last comment? i.e.:

---
core/traffic_controller_rfc2889.py
Line 19:

majority of the code is shared with TrafficControllerRFC2544.__init__(). It would be good to place common code into the ITrafficController and call it from both specialized classes. In this case ITrafficController won't be an abstract class anymore, but the same had happened for other interface classes in the past.

The same logic should be applied for all other methods.
---

If not, then I suggest to create an improvement jira ticket, so it can be addressed in the future.

Best Regards,
Martin

Patch-set: 3