Update patch set 2
Patch Set 2:

Failures of Verify jobs are caused by OVS version used by VSPERF. Vanilla version of this OVS is not compatible with kernel 4.4, which is used on jenkins node used for execution of verify job. Previous (successful) verify jobs were executed at machine with 3.19. It seems, that due to Jenkins reconfiguration only nodes with kernel 4.4. are available.

It means that verify and merge jobs will fail for all patches, until situation will be solved. The best solution seems to OVS tag used by VSPERF as default. This was planned anyway with switch to DPDK 1607 as default version. I'll sent a patch for review shortly.

Patch-set: 2