Update patch set 4
Patch Set 4:

> Hi Otto,
 > 
 > it seems to me, that TRAFFICGEN_STC_RESULTS_DIR is used as a
 > destination directory, where results from STC are copied too. If it
 > is the case, then it should be stored inside VSPERF home directory
 > and not in $PWD. I can also see that default value "./Results" is
 > hardcoded inside following files:
 > 
 > ./tools/pkt_gen/testcenter/testcenter-rfc2544-throughput.py
 > ./tools/pkt_gen/testcenter/testcenter-rfc2544-rest.py
 > 
 > I recommend to contact Sridhar (I've added him to this review) from
 > Spirent and ask him if this code needs to be changed or not. If he
 > confirms that 'Results' subdirectory should be inside VSPERF 'home'
 > dir, then you should do appropriate changes, i.e. to set
 > TRAFFICGEN_STC_RESULTS_DIR inside 03_traffic.conf to
 > os.path.join(ROOT_DIR, 'Results') and change hardcoded defaults
 > from files above to the value of TRAFFICGEN_STC_RESULTS_DIR.
 > 
 > According to the documentation - all examples suppose, that vsperf
 > is executed from its home directory. Also configuration files are
 > sometimes mentioned with their relative paths. It would be good to
 > write inside the doc (e.g. in Executing test section of
 > docs/userguide/testusage.rst) that VSPERF can be executed from any
 > directory, but all examples inside the docs assume, that user is
 > inside the vsperf directory.
 > 
 > Best Regards,
 > Martin

Hello Otto and Martin,

As Mentioned in mail, we do not have to change anything in the source files. Please use the configuration file parameters TRAFFICGEN_STC_RESULTS_DIR and corresponding TRAFFICGEN_STC_CSV_RESULTS_FILE_PREFIX.

This directory is used only to store the CSV file - which will be used to generate the results reports.

As Martin mentioned, if these parameters are not set, the default value for TRAFFICGEN_STC_RESULTS_DIR is ./Results.

The testcenter.py will pass these parameters to the appropriate scripts (for ex: testcenter-rfc2544-rest.py), 
and when the script returns, testcenter.py will pass the .csv file from the TRAFFICGEN_STC_RESULTS_DIR for vsperf to create result reports.

Regards
Sridhar

Patch-set: 4